+++ src/app/lib/app.dart
@@ -1,0 +1,16 @@
+import 'package:app/shell.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+class App extends StatelessWidget {
+  const App({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const ProviderScope(
+      child: MaterialApp(
+        home: Shell(),
+      ),
+    );
+  }
+}