+++ CHANGELOG.md
@@ -14,3 +14,4 @@ Prefix your items with `(Template)` if the change is about the template and not 
 - Target Android 14 with a minimum version of Android 11.
 - Added empty Azure Pipelines.
 - Updated Azure Pipelines (Windows, Android, iOS, Tests).
+- Added simple Dad Jokes page.